create user avatar comp
